Move part of timing module out of the library
