+++ tests/suites/test_suite_timing.data
@@ -1,14 +1,5 @@
-Timing: hardclock
-timing_hardclock:
-
 Timing: get timer
 timing_get_timer:
-
-Timing: set alarm with no delay
-timing_set_alarm:0:
-
-Timing: set alarm with 1s delay
-timing_set_alarm:1:
 
 Timing: delay 0ms
 timing_delay:0:
